pyproject.toml
[project]
name = "webforfileannotation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.14",
    "fitz>=0.0.1.dev2",
    "mysql-connector-python>=9.3.0",
    "pandas>=2.3.0",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.46.1",
    "uvicorn>=0.35.0",
]
